api/pyproject.toml
[project]
name = "medapp-agent"
version = "0.1.0"
description = "An agentic app to schedule medical appointments"
authors = [
    {name = "Carlos Frias",email = "carlos.a.frias@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "faker (>=37.3.0,<38.0.0)",
    "pydantic-extra-types (>=2.10.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
faker = "^37.3.0"

